feat: Move query methods to `QueryUtil` class.
