+++ src/main/java/org/spin/grpc/service/OrderServiceImplementation.java
@@ -27,6 +27,7 @@ import org.compiere.util.Env;
 import org.compiere.util.Util;
 import org.spin.base.db.CountUtil;
 import org.spin.base.db.LimitUtil;
+import org.spin.base.db.QueryUtil;
 import org.spin.base.db.WhereUtil;
 import org.spin.base.util.ContextManager;
 import org.spin.base.util.DictionaryUtil;
@@ -91,7 +92,7 @@ public class OrderServiceImplementation extends OrderImplBase {
 
 		//
 		MTable table = MTable.get(Env.getCtx(), this.tableName);
-		StringBuilder sql = new StringBuilder(DictionaryUtil.getQueryWithReferencesFromColumns(table));
+		StringBuilder sql = new StringBuilder(QueryUtil.getQueryWithReferencesFromTable(table));
 
 		// add where with access restriction
 		String sqlWithRoleAccess = MRole.getDefault()
