+++ src/main/java/org/spin/grpc/service/ProductServiceImplementation.java
@@ -27,6 +27,7 @@ import org.compiere.util.Env;
 import org.compiere.util.Util;
 import org.spin.base.db.CountUtil;
 import org.spin.base.db.LimitUtil;
+import org.spin.base.db.QueryUtil;
 import org.spin.base.db.WhereUtil;
 import org.spin.base.util.ContextManager;
 import org.spin.base.util.DictionaryUtil;
@@ -92,7 +93,7 @@ public class ProductServiceImplementation extends ProductImplBase {
 
 		//
 		MTable table = MTable.get(Env.getCtx(), this.tableName);
-		StringBuilder sql = new StringBuilder(DictionaryUtil.getQueryWithReferencesFromColumns(table));
+		StringBuilder sql = new StringBuilder(QueryUtil.getQueryWithReferencesFromTable(table));
 
 		// add where with access restriction
 		String sqlWithRoleAccess = MRole.getDefault(Env.getCtx(), false)
